// EHealthcare/front/src/app/model/modelartsdiabetes.ts
export class ModelArtsDiabetes{
    highBP: number;
    highChol: number;
    cholCheck: number;
    bmi: number;
    smoker: number;
    stroke: number;
    hearthDisease: number;
    physichalAchivity: number;
    fruit: number;
    veggies: number;
    hvyAlcoholConsump: number;
    anyHealthcare: number;
    noDocbcCost: number;
    genHlth: number;
    mentalHlth: number;
    physHlth: number;
    diffWalk: number;
    sex: number;
    age: number;
    education: number;
    income: number;
}